Starting to update imports and copyrights

+++ megameklab/src/megameklab/ui/largeAero/DSChassisView.java
@@ -1,33 +1,22 @@
 /*
- * MegaMekLab - Copyright (C) 2017 - The MegaMek Team
+ * Copyright (c) 2017-2022 - The MegaMek Team. All Rights Reserved.
  *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 2 of the License, or (at your option) any later
- * version.
+ * This file is part of MegaMekLab.
  *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
+ * MegaMekLab is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MegaMekLab is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MegaMekLab. If not, see <http://www.gnu.org/licenses/>.
  */
 package megameklab.ui.largeAero;
-
-import java.awt.GridBagConstraints;
-import java.awt.GridBagLayout;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.util.List;
-import java.util.ResourceBundle;
-import java.util.concurrent.CopyOnWriteArrayList;
-
-import javax.swing.DefaultComboBoxModel;
-import javax.swing.JCheckBox;
-import javax.swing.JComboBox;
-import javax.swing.JSpinner;
-import javax.swing.SpinnerNumberModel;
-import javax.swing.event.ChangeEvent;
-import javax.swing.event.ChangeListener;
 
 import megamek.common.Dropship;
 import megamek.common.Entity;
@@ -37,6 +26,16 @@ import megamek.common.util.EncodeControl;
 import megamek.common.verifier.TestAero;
 import megameklab.ui.generalUnit.BuildView;
 import megameklab.ui.listeners.DropshipBuildListener;
+
+import javax.swing.*;
+import javax.swing.event.ChangeEvent;
+import javax.swing.event.ChangeListener;
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.util.List;
+import java.util.ResourceBundle;
+import java.util.concurrent.CopyOnWriteArrayList;
 
 /**
  * Structure tab chassis panel for small craft and dropships.
@@ -51,16 +50,16 @@ public class DSChassisView extends BuildView implements ActionListener, ChangeLi
     public void removeListener(DropshipBuildListener l) {
         listeners.remove(l);
     }
-    
+
     public final static int TYPE_SMALL_CRAFT    = 0;
     public final static int TYPE_DROPSHIP       = 1;
-    
+
     public final static int CHASSIS_AERODYNE    = 0;
     public final static int CHASSIS_SPHEROID    = 1;
-    
+
     private final SpinnerNumberModel spnTonnageModel = new SpinnerNumberModel(200, 200, null, 100);
     private final SpinnerNumberModel spnSIModel = new SpinnerNumberModel(1, 1, null, 1);
-    
+
     final private JSpinner spnTonnage = new JSpinner(spnTonnageModel);
     final private JCheckBox chkMilitary = new JCheckBox();
     final private JCheckBox chkKFBoom = new JCheckBox();
@@ -73,12 +72,12 @@ public class DSChassisView extends BuildView implements ActionListener, ChangeLi
     private boolean primitive;
     private int maxTonnage;
     private int maxThrust;
-    
+
     public DSChassisView(ITechManager techManager) {
         this.techManager = techManager;
         initUI();
     }
-    
+
     public void initUI() {
         ResourceBundle resourceMap = ResourceBundle.getBundle("megameklab.resources.Views", new EncodeControl());
         setLayout(new GridBagLayout());
@@ -185,13 +184,13 @@ public class DSChassisView extends BuildView implements ActionListener, ChangeLi
 
         spnSIModel.setValue(craft.get0SI());
     }
-    
+
     public void setAsCustomization() {
         spnTonnage.setEnabled(false);
         cbBaseType.setEnabled(false);
         cbChassisType.setEnabled(false);
     }
-    
+
     public void refresh() {
         refreshTonnage();
         refreshSI();
@@ -256,7 +255,7 @@ public class DSChassisView extends BuildView implements ActionListener, ChangeLi
     }
     
     public void setTonnage(double tonnage) {
-        spnTonnage.setValue(Integer.valueOf((int)Math.ceil(tonnage)));
+        spnTonnage.setValue((int) Math.ceil(tonnage));
     }
     
     public boolean isMilitary() {
@@ -293,5 +292,4 @@ public class DSChassisView extends BuildView implements ActionListener, ChangeLi
             listeners.forEach(l -> l.chassisTypeChanged(cbChassisType.getSelectedIndex()));
         }
     }
-
 }
